Write unit test for warmstarter and fix bug realted to warmstart #22
Warmstarter was not working correctly as the test written for it was failing. Therefore bugs related to warmstart and initial guess predictor bas been fixed in this commit and now the test is passing too.

+++ src/contactstatus.cpp
@@ -4,7 +4,7 @@
 #include <vector>
 
 void ComputeContactNodes(std::vector<double> &xvf, std::vector<double> &yvf,
-    std::vector<double> &pf, int &cont, double &nf, Epetra_SerialDenseMatrix y,
+    std::vector<double> &pf, int &cont, int &nf, Epetra_SerialDenseMatrix y,
     std::vector<double> xv0, std::vector<double> yv0)
 {
   xvf.clear();
@@ -34,7 +34,7 @@ void ComputeContactNodes(std::vector<double> &xvf, std::vector<double> &yvf,
 }
 
 void ComputeContactForceAndArea(std::vector<double> &force0, std::vector<double> &area0, int &iter,
-    double &w_el, double nf, std::vector<double> pf, int k, double delta, double lato, double k_el)
+    double &w_el, int nf, std::vector<double> pf, int k, double delta, double lato, double k_el)
 {
   force0.push_back(0);
   double sum = 0;
